Broke up ProjectViewModels
