+++ src/CompetitionPlatform/Models/ProjectViewModels/CompactStream.cs
@@ -1,0 +1,8 @@
+﻿namespace CompetitionPlatform.Models.ProjectViewModels
+{
+    public class CompactStream
+    {
+        public string  StreamId { get; set; }
+        public string StreamName { get; set; }
+    }
+}
