+++ src/CompetitionPlatform/Models/ProjectViewModels/EditStreamProjects.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace CompetitionPlatform.Models.ProjectViewModels
+{
+    public class EditStreamProjects
+    {
+        public List<StreamProject> ProjectsList { get; set; }
+    }
+}
