+++ src/CompetitionPlatform/Models/ProjectViewModels/LatestWinner.cs
@@ -1,0 +1,11 @@
+﻿namespace CompetitionPlatform.Models.ProjectViewModels
+{
+    public class LatestWinner
+    {
+        public string Name { get; set; }
+        public string ProjectId { get; set; }
+        public string ProjectName { get; set; }
+        public double Amount { get; set; }
+        public string Id { get; set; }
+    }
+}
