+++ src/CompetitionPlatform/Models/ProjectViewModels/ProgrammingResource.cs
@@ -1,0 +1,8 @@
+﻿namespace CompetitionPlatform.Models.ProjectViewModels
+{
+    public class ProgrammingResource
+    {
+        public string Name { get; set; }
+        public string Link { get; set; }
+    }
+}
