+++ src/CompetitionPlatform/Models/ProjectViewModels/ProjectEditStatusDisplayViewModel.cs
@@ -1,0 +1,9 @@
+﻿namespace CompetitionPlatform.Models.ProjectViewModels
+{
+    public class ProjectEditStatusDisplayViewModel
+    {
+        public Status Status { get; set; }
+        public int VotesFor { get; set; }
+        public int VotesAgainst { get; set; }
+    }
+}
