+++ src/CompetitionPlatform/Models/ProjectViewModels/ProjectFileInfoViewModel.cs
@@ -1,0 +1,8 @@
+﻿namespace CompetitionPlatform.Models.ProjectViewModels
+{
+    public class ProjectFileInfoViewModel
+    {
+        public string FileName { get; set; }
+        public string ContentType { get; set; }
+    }
+}
