+++ src/CompetitionPlatform/Models/ProjectViewModels/ProjectParticipateViewModel.cs
@@ -2,19 +2,18 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using CompetitionPlatform.Data.AzureRepositories.Users;
 
 namespace CompetitionPlatform.Models.ProjectViewModels
 {
-    public class ProjectDetailsStatusViewModel
+    public class ProjectParticipateViewModel : IProjectParticipateData
     {
         public string ProjectId { get; set; }
-        public Status Status { get; set; }
-        public int VotesFor { get; set; }
-        public int VotesAgainst { get; set; }
-        public DateTime CompetitionRegistrationDeadline { get; set; }
-        public DateTime ImplementationDeadline { get; set; }
-        public DateTime VotingDeadline { get; set; }
-        public bool IsParticipant { get; set; }
-        public bool HasResult { get; set; }
+        public string UserId { get; set; }
+        public string UserIdentifier { get; set; }
+        public string FullName { get; set; }
+        public DateTime Registered { get; set; }
+        public bool Result { get; set; }
+        public string UserAgent { get; set; }
     }
 }
