+++ src/CompetitionPlatform/Models/ProjectViewModels/ProjectViewModel.cs
@@ -116,100 +116,4 @@ namespace CompetitionPlatform.Models.ProjectViewModels
         [StringLength(30)]
         public string NameTag { get; set; }
     }
-
-    public class ProjectCompactViewModel
-    {
-        public string Id { get; set; }
-        public string Name { get; set; }
-        public string Overview { get; set; }
-        public string Description { get; set; }
-        public Status Status { get; set; }
-        public double BudgetFirstPlace { get; set; }
-        public int VotesFor { get; set; }
-        public int VotesAgainst { get; set; }
-        public int CommentsCount { get; set; }
-        public int ParticipantsCount { get; set; }
-        public int ResultsCount { get; set; }
-        public int WinnersCount { get; set; }
-        public DateTime CompetitionRegistrationDeadline { get; set; }
-        public DateTime ImplementationDeadline { get; set; }
-        public DateTime VotingDeadline { get; set; }
-        public string AuthorFullName { get; set; }
-        public string AuthorId { get; set; }
-        public string Category { get; set; }
-        public List<string> Tags { get; set; }
-        public bool Following { get; set; }
-        public string NameTag { get; set; }
-        public DateTime LastModified { get; set; }
-        public DateTime Created { get; set; }
-    }
-
-    public class ProjectFileInfoViewModel
-    {
-        public string FileName { get; set; }
-        public string ContentType { get; set; }
-    }
-
-    public class ProjectListIndexViewModel
-    {
-        public List<string> ProjectCategories { get; set; }
-        public IEnumerable<ProjectCompactViewModel> Projects { get; set; }
-        public List<LatestWinner> LatestWinners { get; set; }
-        public List<JustFinishedProject> JustFinishedProjects { get; set; }
-    }
-
-    public class ProjectVoteViewModel
-    {
-        public int VotesFor { get; set; }
-        public int VotesAgainst { get; set; }
-    }
-
-    public class ProjectEditStatusDisplayViewModel
-    {
-        public Status Status { get; set; }
-        public int VotesFor { get; set; }
-        public int VotesAgainst { get; set; }
-    }
-
-    public class ProgrammingResource
-    {
-        public string Name { get; set; }
-        public string Link { get; set; }
-    }
-
-    public class LatestWinner
-    {
-        public string Name { get; set; }
-        public string ProjectId { get; set; }
-        public string ProjectName { get; set; }
-        public double Amount { get; set; }
-        public string Id { get; set; }
-    }
-
-    public class JustFinishedProject
-    {
-        public string ProjectId { get; set; }
-        public string ProjectName { get; set; }
-        public double Amount { get; set; }
-        public int NumberOfWinners { get; set; }
-    }
-
-    public class StreamProject
-    {
-        public string ProjectId { get; set; }
-        public string ProjectName { get; set; }
-        public int Place { get; set; }
-    }
-
-    public class CompactStream
-    {
-        public string  StreamId { get; set; }
-        public string StreamName { get; set; }
-    }
-
-    public class EditStreamProjects
-    {
-        public List<StreamProject> ProjectsList { get; set; }
-    }
 }
-
