+++ src/CompetitionPlatform/Models/ProjectViewModels/ProjectVoteViewModel.cs
@@ -1,0 +1,8 @@
+﻿namespace CompetitionPlatform.Models.ProjectViewModels
+{
+    public class ProjectVoteViewModel
+    {
+        public int VotesFor { get; set; }
+        public int VotesAgainst { get; set; }
+    }
+}
