+++ src/CompetitionPlatform/Models/ProjectViewModels/StreamProject.cs
@@ -1,0 +1,9 @@
+﻿namespace CompetitionPlatform.Models.ProjectViewModels
+{
+    public class StreamProject
+    {
+        public string ProjectId { get; set; }
+        public string ProjectName { get; set; }
+        public int Place { get; set; }
+    }
+}
